Independent multipart uploads.
Do completely independent multipart uploads.

First split into separate file
